refactor: changed type imports from types file
