+++ package.json
@@ -21,7 +21,7 @@
   "author": "Braintree <code@getbraintree.com> (https://www.braintreepayments.com/)",
   "license": "MIT",
   "devDependencies": {
-    "@types/jest": "^29.5.2",
+    "@types/jest": "^29.4.0",
     "@types/node": "^18.14.6",
     "@typescript-eslint/eslint-plugin": "^5.54.1",
     "@vitest/coverage-c8": "^0.31.4",
